Complete propertyIsInbetween test
